Fix invalid byte length test
